The blink test was asserting the bug, and /W4 was quietly no longer clean
test_lights had failed two checks for weeks, and it was the TEST that was
wrong. It asserted a 0.667 s period; the source moved to 0.600 deliberately,
because 667 ms is 89.96 flashes per minute and the normally-closed band in
SAE J945 has a floor of 90 - the old "legal standard" figure was just under
the legal minimum. hub/src/lights.hxx carries that reasoning and a matching
static_assert; the test and docs/conventions.md were both still repeating the
number it had replaced.

Now checked in flashes per minute rather than Hz, because that is the unit the
standard is written in and 89.96 against 90 is a distinction Hz hides. The
band itself is asserted alongside the value, so a future change that stays
inside the standard passes and one that leaves it does not.

test_pico_link had not linked at all: pico_link.cxx gained UDP, and inet_pton
needs ws2_32.lib, which build.bat has and the test's link line never got. Ten
unresolved externals, and the whole suite was skipped rather than failing
loudly.

The Array<T, N> conversion cost four /W4 warnings and I did not notice at the
time, because I was grepping the incremental build. sizeof(a) is a compile-time
constant the compiler can prove fits in a DWORD; a.size() is a call it does not
fold the same way, so four ReadFile/fgets sizes started warning about a
narrowing that was always there and always fine. Named casts at all four.
conventions.md claims this tree is warning-clean at /W4 and it is again: a
clean build reports zero.

All eight suites pass and both hardware tests link.

+++ hub/src/recording.cxx
@@ -209,7 +209,7 @@ namespace rec
       }
 
       Array<Char, 64> line= {};
-      if(std::fgets(line.data(), line.size(), f) == nullptr
+      if(std::fgets(line.data(), static_cast<Int32>(line.size()), f) == nullptr
          || (std::strncmp(line.data(), "# biborec", 9) != 0
              && std::strncmp(line.data(), "# tt02rec", 9) != 0))
       {
